refactor(gc): split heap-budget module out of policy.rs (2000-line file lint)
