+++ crates/perry-runtime/src/gc/mod.rs
@@ -38,6 +38,8 @@ pub use types::*;
 mod policy;
 pub(crate) use policy::gc_runtime_safepoint;
 pub use policy::*;
+mod heap_budget;
+pub use heap_budget::*;
 mod telemetry;
 pub use telemetry::*;
 mod malloc;
